==> notebooks/08_Data_Cleaning.ipynb <==
import pandas as pd

df=pd.read_csv("modeling_features.csv")

df.isna().sum()

Many defendants have no prior offenses

df['days_since_last_offense'] = df['days_since_last_offense'].fillna(-1)
df['has_prior_offense'] = (df['days_since_last_offense'] != -1).astype(int)


Categorical identifiers where NaN does NOT mean zero. It means not recorded or not applicable.

categorical_missing = [
    'county','grade_x','subsection','citationcomplaintnumber',
    'citation','title','section','age_group'
]

for col in categorical_missing:
    df[col] = df[col].fillna("Unknown")


Converting to CSV file

df.to_csv("cleaned_modeling_features.csv", index=False)
